w3c-reports/12092014/mac/runned-tests/html/semantics/forms/the-input-element/url-expected.txt

Input url


PASS url type supported on input element 
PASS The value must not be set with "LF" (U+000A) or "CR" (U+000D) 
PASS The value sanitization algorithm is as follows: Strip line breaks from the value 
FAIL The value sanitization algorithm is as follows: Strip leading and trailing whitespace from the value. assert_equals: expected "aa" but got " aa "(stack: assert@http://localhost:8800/resources/testharness.js:1738:42
assert_equals@http://localhost:8800/resources/testharness.js:335:15
http://localhost:8800/html/semantics/forms/the-input-element/url.html:51:17
step@http://localhost:8800/resources/testharness.js:798:30
test@http://localhost:8800/resources/testharness.js:73:22
global code@http://localhost:8800/html/semantics/forms/the-input-element/url.html:49:7)
